fix(wsi): Remove PydicomHandler dependency from dcm file scanning
